test(color-picker): recategorize and clean up browser tests
